refactor: MemberService Architecture (#6)
