+++ woorizip-core/src/main/kotlin/core/member/service/MemberService.kt
@@ -3,7 +3,9 @@ package core.member.service
 class MemberService(
     queryMemberService: QueryMemberService,
     commandMemberService: CommandMemberService,
+    commandMemberExternalService: CommandMemberExternalService,
     checkMemberService: CheckMemberService
 ) : QueryMemberService by queryMemberService,
     CommandMemberService by commandMemberService,
+    CommandMemberExternalService by commandMemberExternalService,
     CheckMemberService by checkMemberService
